+++ lib/actionize/extensions/converting_null_param_values_to_nil.rb
@@ -1,15 +1,13 @@
 # Converts the following filter request param values to nil: 'NULL', 'null', 'nil'
-module RestfulJson
-  module Controller
+module Actionize
+  module Extensions
     module ConvertingNullParamValuesToNil
       extend ::ActiveSupport::Concern
-
-      include ::RestfulJson::Controller
 
       NILS = ['NULL'.freeze, 'null'.freeze, 'nil'.freeze].to_set
 
       # Converts request param values 'NULL', 'null', and 'nil' to nil.
-      def convert_request_param_value_for_filtering(attr_sym, param_value)
+      def convert_param_value(param_name, param_value)
         param_value = super || param_value
         param_value && NILS.include?(param_value) ? nil : param_value
       end