+++ lib/actionize/extensions/using_standard_rest_render_options.rb
@@ -1,19 +1,17 @@
 # Sets standard status and location properly in create and update in rendering options.
-module RestfulJson
-  module Controller
+module Actionize
+  module Extensions
     module UsingStandardRestRenderOptions
       extend ::ActiveSupport::Concern
 
-      include ::RestfulJson::Controller
-
       def render_create_valid_options(value)
-        (super || {}).merge!({status: :created, location: value})
+        (super || {}).merge!({status: :created})
       end
 
       def render_update_valid_options(value)
         (super || {}).merge!({status: :no_content})
       end
-
+      
     end
   end
 end